Feat : PDFPreviewModal에서 CategoryChip, SeasonChip 사용
